requirements.txt
discord.py[voice]
google-generativeai
jupyterlab
pymongo